+++ common/src/main/java/me/rufia/fightorflight/goals/targeting/PokemonNearestAttackableTargetGoal.java
@@ -9,7 +9,7 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
 
-
+@Deprecated
 public class PokemonNearestAttackableTargetGoal<T extends LivingEntity> extends NearestAttackableTargetGoal<T> {
     public int ticksUntilNewAngerParticle = 0;
     public float safeDistanceSqr = 36;